refactor: AreaResponseDto 위치 이동 import 수정
