+++ src/main/java/com/nineteen/omp/store/repository/AreaRepositoryCustom.java
@@ -1,6 +1,6 @@
 package com.nineteen.omp.store.repository;
 
-import com.nineteen.omp.store.service.dto.AreaResponseDto;
+import com.nineteen.omp.store.controller.dto.AreaResponseDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
